+++ arez/src/main/java/react4j/arez/ReactArezComponent.java
@@ -15,6 +15,7 @@ import elemental2.core.JsObject;
 import java.util.List;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import javax.annotation.OverridingMethodsMustInvokeSuper;
 import jsinterop.base.Js;
 import jsinterop.base.JsPropertyMap;
 import org.realityforge.braincheck.Guards;
@@ -166,6 +167,7 @@ public abstract class ReactArezComponent
    * {@inheritDoc}
    */
   @Override
+  @OverridingMethodsMustInvokeSuper
   protected boolean shouldComponentUpdate( @Nullable final JsPropertyMap<Object> nextProps,
                                            @Nullable final JsPropertyMap<Object> nextState )
   {
@@ -199,6 +201,7 @@ public abstract class ReactArezComponent
    * {@inheritDoc}
    */
   @Override
+  @OverridingMethodsMustInvokeSuper
   protected void componentDidMount()
   {
     storeArezDataAsState();
@@ -208,6 +211,7 @@ public abstract class ReactArezComponent
    * {@inheritDoc}
    */
   @Override
+  @OverridingMethodsMustInvokeSuper
   protected void componentDidUpdate( @Nullable final JsPropertyMap<Object> prevProps,
                                      @Nullable final JsPropertyMap<Object> prevState )
   {
@@ -218,6 +222,7 @@ public abstract class ReactArezComponent
    * {@inheritDoc}
    */
   @Override
+  @OverridingMethodsMustInvokeSuper
   protected void componentWillUnmount()
   {
     /*